Implementing next/prev Buttons

+++ todo-react-app/src/Components/ToDo.js
@@ -1,18 +1,25 @@
 import React, { useEffect, useState, useContext } from "react";
 import Form from "./Form";
-import "./navBar.css";
+import "./components.css";
 import { Card, H5, Elevation } from "@blueprintjs/core";
 import { SettingsContext } from "../Context/settings";
 
 export default function ToDo(props) {
   const settings = useContext(SettingsContext);
   const [list, setList] = useState([]);
+  const [index, setIndex] = useState(0);
   const [incomplete, setIncomplete] = useState([]);
 
   function deleteItem(id) {
     const items = list.filter((item) => item.id !== id);
     setList(items);
   }
+
+  let listParts = list.reduce((all, one, i) => {
+    const ch = Math.floor(i / settings.itemsNumber);
+    all[ch] = [].concat(all[ch] || [], one);
+    return all;
+  }, []);
 
   function toggleComplete(id) {
     const items = list.map((item) => {
@@ -31,34 +38,45 @@ export default function ToDo(props) {
     document.title = `To Do List: ${incomplete}`;
   }, [list]);
 
-  return (
-    <>
-      <Form list={list} incomplete={incomplete} setList={setList} />
-      <div className="listContainer">
-        {list.map((item) => (
-          <div key={item.id} className="cards">
-            {/* if(!(item.complete && !(settungs.showCompleted))){ */}
-            <Card elevation={Elevation.THREE}>
-              <H5>{item.text}</H5>
-              <p>Assigned to: {item.assignee}</p>
-              <div onClick={() => toggleComplete(item.id)}>
-                Complete: {item.complete.toString()}
+  if (listParts[index]) {
+    return (
+      <>
+        <Form list={list} incomplete={incomplete} setList={setList} />
+        <div className="listContainer">
+          {/* Next Button */}
+          {list.length > settings.itemsNumber &&
+          index < listParts.length - 1 ? (
+            <button id='nextBtn' className='traversButtons' onClick={() => setIndex(index + 1)}>Next</button>
+          ) : (
+            <div />
+          )}
+          {/* Prev Button */}
+          {index > 0 ? (
+            <button className='traversButtons' onClick={() => setIndex(index - 1)}>Prev</button>
+          ) : (
+            <div />
+          )}
+          {listParts[index].map((item) => {
+            return (
+              <div key={item.id} className="cards">
+                {/* if(!(item.complete && !(settungs.showCompleted))){ */}
+                <Card elevation={Elevation.THREE}>
+                  <H5>{item.text}</H5>
+                  <p>Assigned to: {item.assignee}</p>
+                  <div onClick={() => toggleComplete(item.id)}>
+                    Complete: {item.complete.toString()}
+                  </div>
+                  <button onClick={() => deleteItem(item.id)}>Close</button>
+                </Card>
+                {/* }} */}
+                <hr />
               </div>
-              <button onClick={() => deleteItem(item.id)}>Close</button>
-            </Card>
-          {/* }} */}
-            <hr />
-          </div>
-        ))}
-        {
-          // list.length > settings.itemsNumber ? <button>Next</button> : {}
-        }
-      </div>
-    </>
-  );
+            );
+          })}
+        </div>
+      </>
+    );
+  } else {
+    return <Form list={list} incomplete={incomplete} setList={setList} />;
+  }
 }
-
-// complete : true // settings : false // no
-// complete : true // settings : true // yes
-// complete : false // settings : false // yes
-// complete : false // settings : true // yes
